requirements.txt
alembic==0.9.9
attrs==18.1.0
Babel==2.8.0
bcrypt==3.1.4
bleach==3.1.3
blinker==1.4
cffi==1.11.5
Click==7.0
coverage==4.5.1
cryptography==2.8
dnspython==1.16.0
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.9.0
Flask-MonitoringDashboard==3.0.9
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
git+git://github.com/maxcountryman/flask-uploads@f66d7dc
Flask-WTF==0.14.3
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.0.7
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.1.0
numpy==1.18.2
passlib==1.7.1
pluggy==0.13.1
psycogreen==1.0.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.5.3
pycparser==2.18
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.8.2
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2018.4
redis==3.4.1
rq==1.3.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.15
webencodings==0.5.1
Werkzeug==1.0.0
WTForms==2.1
